Add shimmer package for loading effects, update pubspec.yaml, and enhance payment method dialogs with new payment handling features. Refactor MoosylView and payment methods to streamline user experience and improve code organization.
